builds/requirements.txt
pyoxidizer==0.20.0
PyYAML==6.0
